+++ src/Farmer/Builders/Builders.ContainerGroups.fs
@@ -72,6 +72,8 @@ type InitContainerConfig =
 type ContainerGroupConfig =
     { /// The name of the container group.
       Name : ResourceName
+      /// DNS configuration for the container group
+      DnsConfig : ContainerGroupDnsConfiguration option
       /// Container group OS.
       OperatingSystem : OS
       /// Restart policy for the container group.
@@ -118,6 +120,10 @@ type ContainerGroupConfig =
                            ReadinessProbe = instance.ReadinessProbe
                            VolumeMounts = instance.VolumeMounts |}
               ]
+              DnsConfig =
+                  if this.DnsConfig.IsSome && this.NetworkProfile.IsNone then
+                    raiseFarmer "DNS configuration can only be set when attached to a virtual network."
+                  else this.DnsConfig
               OperatingSystem = this.OperatingSystem
               RestartPolicy = this.RestartPolicy
               Identity = this.Identity
@@ -190,6 +196,7 @@ type ContainerGroupBuilder() =
 
     member _.Yield _ =
         { Name = ResourceName.Empty
+          DnsConfig = None
           OperatingSystem = Linux
           RestartPolicy = AlwaysRestart
           Identity = ManagedIdentity.Empty
@@ -258,6 +265,42 @@ type ContainerGroupBuilder() =
         let newVolumes = volumes |> Map.ofSeq
         let updatedVolumes = state.Volumes |> Map.fold (fun current key vol -> Map.add key vol current) newVolumes
         { state with Volumes = updatedVolumes }
+    /// Specify DNS nameservers for the containers in the container group.
+    [<CustomOperation "dns_nameservers">]
+    member _.DnsNameServers(state:ContainerGroupConfig, nameServers:string list) =
+        let dns =
+            match state.DnsConfig with
+            | None ->
+                { NameServers = nameServers
+                  Options = []
+                  SearchDomains = [] }
+            | Some dnsConfig ->
+                { dnsConfig with NameServers = nameServers }
+        { state with DnsConfig = Some dns }
+    /// Specify DNS options (e.g. 'ndots:2') for the containers in the container group.
+    [<CustomOperation "dns_options">]
+    member _.DnsOptions(state:ContainerGroupConfig, options:string list) =
+        let dns =
+            match state.DnsConfig with
+            | None ->
+                { NameServers = []
+                  Options = options
+                  SearchDomains = [] }
+            | Some dnsConfig ->
+                { dnsConfig with Options = options }
+        { state with DnsConfig = Some dns }
+    /// Specify DNS search domains for the containers in the container group.
+    [<CustomOperation "dns_search_domains">]
+    member _.DnsSearchDomains(state:ContainerGroupConfig, searchDomains:string list) =
+        let dns =
+            match state.DnsConfig with
+            | None ->
+                { NameServers = []
+                  Options = []
+                  SearchDomains = searchDomains }
+            | Some dnsConfig ->
+                { dnsConfig with SearchDomains = searchDomains }
+        { state with DnsConfig = Some dns }
 
     interface IIdentity<ContainerGroupConfig> with member _.Add state updater = { state with Identity = updater state.Identity }
     interface ITaggable<ContainerGroupConfig> with member _.Add state tags = { state with Tags = state.Tags |> Map.merge tags }